requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
openpyxl>=3.1.0
plotly>=5.18.0
msoffcrypto-tool>=5.0.0
PyPDF2>=3.0.0
psycopg2-binary>=2.9.0
